feat: adiciona UUID como identificador único de tarefas e reorganiza estrutura em arquivos múltiplus
