+++ data/db.json
@@ -1,12 +1,7 @@
 [
     {
-        "id": 2,
-        "titulo": "CRUD-python",
-        "concluida": true
-    },
-    {
-        "id": 3,
-        "titulo": "qual ID essa tarefa ter\u00e1?",
+        "id": "85f79251-f4e8-4401-bac6-224456c34a92",
+        "titulo": "Testando UUID",
         "concluida": false
     }
 ]
